Refactor TSV parsing logic to shared module
Moved the parseTSVWithQuotes function to a new handleCopy.js utility for reuse. Updated LapusFenomena.vue and Create.vue to import and use the shared function. Added paste handling in Create.vue to support pasting TSV data into the rows selection.
